+++ blocklets/image-bin/.eslintrc.js
@@ -11,4 +11,12 @@ module.exports = {
   globals: {
     logger: true,
   },
+  rules: {
+    'react/require-default-props': [
+      'error',
+      {
+        functions: 'defaultArguments',
+      },
+    ],
+  },
 };