+++ blocklets/image-bin/src/components/layout.jsx
@@ -17,7 +17,7 @@ import { useUploadContext } from '../contexts/upload';
 import { useSessionContext } from '../contexts/session';
 import { hasAdminPermission, hasMediaKitAccessPermission } from '../libs/utils';
 
-export default function Layout({ title }) {
+export default function Layout({ title = window.blocklet.appName }) {
   const { tab } = useUploadContext();
   const { session } = useSessionContext();
   const navigate = useNavigate();
@@ -40,7 +40,8 @@ export default function Layout({ title }) {
 
   const DashWrapper = useMemo(() => {
     return adminPermissionInSingleTenant
-      ? StyledDashboard
+      ? // eslint-disable-next-line no-use-before-define
+        StyledDashboard
       : ({ children }) => (
           <Box sx={{ height: '100vh', overflowY: 'hidden' }}>
             <Header
@@ -143,8 +144,4 @@ const StyledDashboard = styled(Dashboard)`
 
 Layout.propTypes = {
   title: PropTypes.string,
-};
-
-Layout.defaultProps = {
-  title: window.blocklet.appName,
 };